% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acronymize.R
\name{Acronymize}
\alias{Acronymize}
\title{Acronymize}
\usage{
Acronymize(phrase, separator)
}
\arguments{
\item{phrase}{This is a string value that will turned into an acronym}

\item{separator}{This can either be "" or .}
}
\value{
Returns an acronym of the input string
}
\description{
This function turns a phrase into an acronym
}
\examples{
Acronymize("International Monetary Fund", separator = "")
}
