% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anagram.R
\name{anagramChecker}
\alias{anagramChecker}
\title{Anagram Checker}
\usage{
anagramChecker(word1, word2)
}
\arguments{
\item{word1}{This is the first word to be checked against the second word}

\item{word2}{This is the second word to be checked against the first word}
}
\value{
Returns a boolean value
}
\description{
This function checks if two words are anagrams of each other
}
\examples{
anagramChecker("tight","light")
}
