% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capitalize.R
\name{Capitalize}
\alias{Capitalize}
\title{Capitalize}
\usage{
Capitalize(string)
}
\arguments{
\item{string}{This any word passed to the function}
}
\value{
Returns a capitalized version of the word passed to the function
}
\description{
This function capitalizes any word passed to it
}
\examples{
Capitalize("institution")
}
