% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isacronym.R
\name{is.acronym}
\alias{is.acronym}
\title{Is Acronym}
\usage{
is.acronym(string)
}
\arguments{
\item{string}{This is any string value or word which the function checks}
}
\value{
Returns a boolean value
}
\description{
This function checks whether or not a string is an acronym
}
\examples{
is.acronym("UNICEF")
}
