% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscapitalized.R
\name{is.capitalized}
\alias{is.capitalized}
\title{Is Capitalized}
\usage{
is.capitalized(string)
}
\arguments{
\item{string}{This is any string value which the function checks for evidence of capitalization}
}
\value{
Returns a boolean value
}
\description{
This function checks whether or not a string is capitalized
}
\examples{
is.capitalized("message")
}
