% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islower.R
\name{is.lowercase}
\alias{is.lowercase}
\title{Is Lower}
\usage{
is.lowercase(string)
}
\arguments{
\item{string}{This is any string value which the function checks}
}
\value{
Returns a boolean value
}
\description{
This function checks whether or not a string is in lowercase characters
}
\examples{
is.lower("maximum")
}
