% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispalindrome.R
\name{is.palindrome}
\alias{is.palindrome}
\title{Is Palindrome}
\usage{
is.palindrome(input)
}
\arguments{
\item{input}{This can either be a string or numeric value}
}
\value{
Returns a boolean value
}
\description{
This function checks whether or not a string or numeric value is a palindrome
}
\examples{
is.palindrome(858)
is.palindrome("Bot")
}
