% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isupper.R
\name{is.uppercase}
\alias{is.uppercase}
\title{Is Upper}
\usage{
is.uppercase(string)
}
\arguments{
\item{string}{This is any string value which the function checks}
}
\value{
Returns a boolean value
}
\description{
This function checks whether or not a string is in uppercase characters
}
\examples{
is.upper("FACTUAL")
}
