% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalindromeExt.R
\name{palindromeExtract}
\alias{palindromeExtract}
\title{Palindrome Extract}
\usage{
palindromeExtract(input)
}
\arguments{
\item{input}{This can either be a sentence or a vector of string or numeric values}
}
\value{
Returns the palindromes in the input data
}
\description{
This function checks for and returns any palindromes in a sentence or vector
}
\examples{
palindromeExtract(c("Dad","Toy","Mum","Car"))
}
