% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrange.R
\name{rearrangeWays}
\alias{rearrangeWays}
\title{Rearrange}
\usage{
rearrangeWays(string, as_report = TRUE)
}
\arguments{
\item{string}{This any string value passed to the function to be rearranged}

\item{as_report}{This takes a boolean value. If TRUE, the function prints a short report in the console. If FALSE, it prints out the number of possible ways in which a string value can be rearranged. It defaults to TRUE.}
}
\value{
Returns either a numeric value or short report.
}
\description{
This function returns the number of possible ways in which a word or string can be rearranged
}
\examples{
rearrangeWays("tension",as_report = FALSE)
}
