% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortenVector.R
\name{shorten.vector}
\alias{shorten.vector}
\title{Shorten Vector}
\usage{
shorten.vector(vector, by)
}
\arguments{
\item{vector}{This a vector of string or numeric values}

\item{by}{This is by how many values the vector should be shortened}
}
\value{
Returns either a string vector or a numeric vector
}
\description{
This function shortens the length of vector
}
\examples{
shorten.vector(c("apple","pear","grape","peach"), by = 2)
}
